+++ diagnostic_interface/__init__.py
@@ -3,11 +3,13 @@ from .settings_window import SettingsDialog
 from .custom_toolbar import CustomNavigationToolbar
 from .plot_canvas import PlotCanvas
 from .plot_tab import PlotTab
+from .data_select import DataSelect
 
 __all__ = [
     "TimedWidget",
     "SettingsDialog",
     "CustomNavigationToolbar",
     "PlotCanvas",
-    "PlotTab"
+    "PlotTab",
+    "DataSelect"
 ]
